# frontend/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DailyTools247 API",
    "description": "Comprehensive set of free utility and AI-powered APIs for developers, including text analysis, security checkers, formatting utilities, social media copy generation, and email marketing analytics.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://www.dailytools247.app/api",
      "description": "Production Server"
    }
  ],
  "paths": {
...
    "/dev/json-formatter": {
      "post": {
        "summary": "Format, validate, or minify JSON data",
        "description": "Cleans up and beautifies raw JSON strings, minifies them, or performs format validation.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["json"],
                "properties": {
                  "json": { "type": "string" },
                  "action": {
                    "type": "string",
                    "enum": ["format", "minify", "validate"],
                    "default": "format"
                  },
                  "indent": { "type": "integer", "default": 2 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "formatted": { "type": "string" },
                        "valid": { "type": "boolean" },
                        "error": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dev/url-encoder": {
      "post": {
        "summary": "URL encode or decode string values",
        "description": "Safely encodes query parameters or decodes URL components using standard percent-encoding.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": { "type": "string" },
                  "action": {
                    "type": "string",
                    "enum": ["encode", "decode"],
                    "default": "encode"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "output": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dev/ai-cron-generator": {
      "post": {
        "summary": "Generate cron expressions from natural language description",
        "description": "Uses AI to convert plain text descriptions of schedules (e.g. 'every Monday at 5 PM') into valid cron expressions.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["description"],
                "properties": {
                  "description": { "type": "string", "description": "e.g., 'every Monday at 5 PM'" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "cron": { "type": "string" },
                        "explanation": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dev/ai-json-to-typescript-interface": {
      "post": {
        "summary": "Convert JSON structure into TypeScript interfaces",
        "description": "Translates arbitrary JSON input into strongly typed TypeScript interface code blocks dynamically using AI.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["json"],
                "properties": {
                  "json": { "type": "string" },
                  "rootName": { "type": "string", "default": "RootInterface" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "typescript": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dev/ai-sql-query-beautifier": {
      "post": {
        "summary": "Beautify, lint, and format SQL queries",
        "description": "Processes raw SQL text and reformats it with standardized keywords and indentations.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sql"],
                "properties": {
                  "sql": { "type": "string" },
                  "dialect": { "type": "string", "default": "sql" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "formattedSql": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}